backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.22

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0

# Security
cryptography==46.0.5
PyJWT[crypto]==2.11.0
bcrypt>=4.0.0

# Background jobs
apscheduler==3.10.4

# Async support
aiofiles==24.1.0
httpx==0.28.1

# parsedmarc and dependencies (already installed in this project)
# parsedmarc will be available from the parent directory

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Rate limiting
slowapi==0.1.9

# Utils
python-dotenv==1.0.1
